(PC-22797)[PRO] refactor: Refactor `ico-pen` usage to fix a size issue.
