+++ pro/src/icons/index.ts
@@ -31,7 +31,6 @@ export { ReactComponent as OfferCardEuroIcon } from 'icons/offer-card-euro.svg'
 export { ReactComponent as OtherOfferIcon } from 'icons/other-offer.svg'
 export { ReactComponent as PartyIcon } from 'icons/party.svg'
 export { ReactComponent as PasscultureIcon } from 'icons/ico-passculture.svg'
-export { ReactComponent as PenIcon } from 'icons/ico-pen.svg'
 export { ReactComponent as PlusCircleIcon } from 'icons/ico-plus-circle.svg'
 export { ReactComponent as ResetIcon } from 'icons/reset.svg'
 export { ReactComponent as SearchIcon } from 'icons/ico-search.svg'
